requirements.txt
requests>=2.31.0
openai>=1.30.0
click>=8.1.7
rich>=13.7.0
python-dotenv>=1.0.0
beautifulsoup4>=4.12.0
